DB Update: March 20th
This should be the (mostly) final version of the DB program. There is still some test code left over (such as the print statements within the functions that should be removed and the temporary main function with different test scenarios), but all of the implementations we need based off of our design diagram should be there.